[SYCL][RTC] Make JIT-based pipeline the default for SYCL language (#17383)
Remove the invoke-based implementation and the temporary
`source_language::sycl_jit` - the in-memory pipeline now handles
`source_language::sycl`.

---------
