+++ sycl/include/sycl/kernel_bundle_enums.hpp
@@ -25,7 +25,6 @@ enum class source_language : int {
   spirv = 1,
   sycl = 2,
   /* cuda */
-  sycl_jit = 99 /* temporary, alternative implementation for SYCL */
 };
 
 // opencl versions
