+++ sycl/source/kernel_bundle.cpp
@@ -370,8 +370,6 @@ bool is_source_kernel_bundle_supported(backend BE, source_language Language) {
     } else if (Language == source_language::spirv) {
       return true;
     } else if (Language == source_language::sycl) {
-      return detail::SYCL_Compilation_Available();
-    } else if (Language == source_language::sycl_jit) {
       return detail::SYCL_JIT_Compilation_Available();
     }
   }
