+++ sycl/test-e2e/KernelCompiler/opencl_capabilities.cpp
@@ -1,4 +1,4 @@
-//==- opencl_capabilities.cpp ----------------------------------------------==//
+//==--- opencl_capabilities.cpp --- kernel_compiler extension tests --------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
