+++ sycl/test-e2e/KernelCompiler/opencl_local_mem.cpp
@@ -1,3 +1,11 @@
+//==--- opencl_local_mem.cpp --- kernel_compiler extension tests -----------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
 // Tests using an OpenCL-C defined kernel with multiple local accessors
 
 // REQUIRES: ocloc && (opencl || level_zero)
