+++ sycl/test-e2e/KernelCompiler/opencl_multi_device.cpp
@@ -1,3 +1,11 @@
+//==--- opencl_multi_device.cpp --- kernel_compiler extension tests --------==//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
 // REQUIRES: (opencl || level_zero) && ocloc
 // UNSUPPORTED: accelerator
 // UNSUPPORTED-INTENDED: while accelerator is AoT only, this cannot run there.
