+++ sycl/test-e2e/KernelCompiler/opencl_queries_negative.cpp
@@ -1,4 +1,4 @@
-//==------- ocloc_queries_negative.cpp -------------------------------------==//
+//==--- opencl_queries_negative.cpp --- kernel_compiler extension tests ----==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
