+++ sycl/test-e2e/KernelCompiler/spirv.cpp
@@ -1,4 +1,4 @@
-//==- kernel_compiler_spirv.cpp --------------------------------------------==//
+//==--- spirv.cpp --- kernel_compiler extension tests --------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
