+++ sycl/test-e2e/KernelCompiler/sycl_basic.cpp
@@ -1,4 +1,4 @@
-//==- kernel_compiler_overload.cpp --- kernel_compiler extension tests -----==//
+//==--- sycl_basic.cpp --- kernel_compiler extension tests -----------------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -18,6 +18,7 @@
 #include <sycl/detail/core.hpp>
 #include <sycl/kernel_bundle.hpp>
 #include <sycl/usm.hpp>
+
 namespace syclexp = sycl::ext::oneapi::experimental;
 
 static constexpr size_t NUM = 1024;
@@ -26,45 +27,37 @@ static constexpr size_t WGSIZE = 16;
 int main() {
   sycl::queue q;
 
-  // The source code for two kernels defined as overloaded functions.
+  // The source code for a kernel, defined as a SYCL "free function kernel".
   std::string source = R"""(
     #include <sycl/sycl.hpp>
     namespace syclext = sycl::ext::oneapi;
     namespace syclexp = sycl::ext::oneapi::experimental;
 
+    extern "C"
     SYCL_EXT_ONEAPI_FUNCTION_PROPERTY((syclexp::nd_range_kernel<1>))
     void iota(float start, float *ptr) {
       size_t id = syclext::this_work_item::get_nd_item<1>().get_global_linear_id();
       ptr[id] = start + static_cast<float>(id);
-    }
-
-    SYCL_EXT_ONEAPI_FUNCTION_PROPERTY((syclexp::nd_range_kernel<1>))
-    void iota(int start, int *ptr) {
-      size_t id = syclext::this_work_item::get_nd_item<1>().get_global_linear_id();
-      ptr[id] = start + static_cast<int>(id);
     }
   )""";
 
   // Create a kernel bundle in "source" state.
   sycl::kernel_bundle<sycl::bundle_state::ext_oneapi_source> kb_src =
       syclexp::create_kernel_bundle_from_source(
-          q.get_context(), syclexp::source_language::sycl_jit, source);
+          q.get_context(), syclexp::source_language::sycl, source);
 
-  // Compile the kernel.  Because there are two overloads of "iota", we need to
-  // use a C++ cast to disambiguate between them.  Here, we are selecting the
-  // "int" overload.
-  std::string iota_name{"(void(*)(int, int*))iota"};
-  sycl::kernel_bundle<sycl::bundle_state::executable> kb_exe = syclexp::build(
-      kb_src, syclexp::properties{syclexp::registered_names{iota_name}});
+  // Compile the kernel.  There is no need to use the "registered_names"
+  // property because the kernel is declared extern "C".
+  sycl::kernel_bundle<sycl::bundle_state::executable> kb_exe =
+      syclexp::build(kb_src);
 
-  // Get the kernel by passing the same string we used to construct the
-  // "registered_names" property.
-  sycl::kernel iota = kb_exe.ext_oneapi_get_kernel(iota_name);
+  // Get the kernel via its compiler-generated name.
+  sycl::kernel iota = kb_exe.ext_oneapi_get_kernel("iota");
 
-  int *ptr = sycl::malloc_shared<int>(NUM, q);
+  float *ptr = sycl::malloc_shared<float>(NUM, q);
   q.submit([&](sycl::handler &cgh) {
      // Set the values of the kernel arguments.
-     cgh.set_args(3, ptr);
+     cgh.set_args(3.14f, ptr);
 
      // Launch the kernel according to its type, in this case an nd-range
      // kernel.
@@ -72,8 +65,10 @@ int main() {
      cgh.parallel_for(ndr, iota);
    }).wait();
 
+  constexpr float eps = 0.001;
   for (int i = 0; i < NUM; i++) {
-    if (ptr[i] != i + 3) {
+    const float truth = 3.14f + static_cast<float>(i);
+    if (std::abs(ptr[i] - truth) > eps) {
       std::cout << "Result: " << ptr[i] << " expected " << i << "\n";
       sycl::free(ptr, q);
       exit(1);
